Refactor StreamConfig usage and add JS types
- Remove StreamConfig duplication from scan/streaming.go
- Use config.StreamConfig directly in js/streaming.go
- Create internal/lang/js/types.go with ImportSpec, ExportSymbol, FlagOccurrence
- Remove duplicate type definitions from collector.go
